requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.11.3'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
crispy-bootstrap4==2024.1; python_version >= '3.8'
django==4.2.7; python_version >= '3.8'
django-allauth==0.63.3; python_version >= '3.7'
django-appconf==1.0.6; python_version >= '3.7'
django-crispy-forms==2.1; python_version >= '3.8'
django-fsm==3.0.0
django-imagekit==5.0.0
django-service-objects==0.7.1
django-stubs==5.0.2; python_version >= '3.8'
django-stubs-ext==5.0.2; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
kombu==5.3.7; python_version >= '3.8'
pilkit==3.0
pillow==10.3.0; python_version >= '3.8'
prompt-toolkit==3.0.46; python_full_version >= '3.7.0'
psycopg==3.1.19; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
redis==5.0.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
types-pyyaml==6.0.12.20240311; python_version >= '3.8'
typing-extensions==4.12.1; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uuid==1.30
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13